Move plugin settings into settings.py files
Also, use a consistent naming for plugin settings and ensure every plugin
setting has a default and a validator.
