+++ plugins/google_analytics/girder_google_analytics/__init__.py
@@ -1,15 +1,7 @@
 # -*- coding: utf-8 -*-
-from girder.exceptions import ValidationException
 from girder.plugin import GirderPlugin
-from girder.utility import setting_utilities
 
-from . import constants, rest
-
-
-@setting_utilities.validator(constants.PluginSettings.GOOGLE_ANALYTICS_TRACKING_ID)
-def validateTrackingId(doc):
-    if not doc['value']:
-        raise ValidationException('Google Analytics Tracking ID must not be empty.', 'value')
+from . import rest
 
 
 class GoogleAnalyticsPlugin(GirderPlugin):
