--- plugins/google_analytics/girder_google_analytics/constants.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-# Constants representing the setting keys for this plugin
-class PluginSettings(object):
-    GOOGLE_ANALYTICS_TRACKING_ID = 'google_analytics.tracking_id'
