+++ plugins/google_analytics/girder_google_analytics/rest.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
-from . import constants
 from girder.api import access
 from girder.api.describe import Description, describeRoute
 from girder.api.rest import Resource
 from girder.models.setting import Setting
+
+from .settings import PluginSettings
 
 
 class GoogleAnalytics(Resource):
@@ -17,5 +18,5 @@ class GoogleAnalytics(Resource):
         Description('Public url for getting the Google Analytics tracking id.')
     )
     def getId(self, params):
-        trackingId = Setting().get(constants.PluginSettings.GOOGLE_ANALYTICS_TRACKING_ID)
+        trackingId = Setting().get(PluginSettings.TRACKING_ID)
         return {'google_analytics_id': trackingId}
