+++ plugins/gravatar/girder_gravatar/settings.py
@@ -1,0 +1,16 @@
+from girder.utility import setting_utilities
+
+
+class PluginSettings(object):
+    DEFAULT_IMAGE = 'gravatar.default_image'
+
+
+@setting_utilities.default(PluginSettings.DEFAULT_IMAGE)
+def _defaultDefaultImage():
+    return 'identicon'
+
+
+@setting_utilities.validator(PluginSettings.DEFAULT_IMAGE)
+def _validateDefaultImage(doc):
+    # TODO should we update user collection to remove gravatar_baseUrl vals?
+    pass
