--- plugins/ldap/girder_ldap/constants.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-class PluginSettings(object):
-    LDAP_SERVERS = 'ldap.servers'
