+++ plugins/oauth/girder_oauth/__init__.py
@@ -4,39 +4,8 @@ from girder.constants import SortDir
 from girder.exceptions import ValidationException
 from girder.models.user import User
 from girder.plugin import GirderPlugin
-from girder.settings import SettingDefault
-from girder.utility import setting_utilities
-from . import rest, constants, providers
 
-
-@setting_utilities.validator(constants.PluginSettings.PROVIDERS_ENABLED)
-def validateProvidersEnabled(doc):
-    if not isinstance(doc['value'], (list, tuple)):
-        raise ValidationException('The enabled providers must be a list.', 'value')
-
-
-@setting_utilities.validator(constants.PluginSettings.IGNORE_REGISTRATION_POLICY)
-def validateIgnoreRegistrationPolicy(doc):
-    if not isinstance(doc['value'], bool):
-        raise ValidationException('Ignore registration policy setting must be boolean.', 'value')
-
-
-@setting_utilities.validator({
-    constants.PluginSettings.GOOGLE_CLIENT_ID,
-    constants.PluginSettings.GLOBUS_CLIENT_ID,
-    constants.PluginSettings.GITHUB_CLIENT_ID,
-    constants.PluginSettings.LINKEDIN_CLIENT_ID,
-    constants.PluginSettings.BITBUCKET_CLIENT_ID,
-    constants.PluginSettings.BOX_CLIENT_ID,
-    constants.PluginSettings.GOOGLE_CLIENT_SECRET,
-    constants.PluginSettings.GLOBUS_CLIENT_SECRET,
-    constants.PluginSettings.GITHUB_CLIENT_SECRET,
-    constants.PluginSettings.LINKEDIN_CLIENT_SECRET,
-    constants.PluginSettings.BITBUCKET_CLIENT_SECRET,
-    constants.PluginSettings.BOX_CLIENT_SECRET
-})
-def validateOtherSettings(event):
-    pass
+from . import rest, providers
 
 
 def checkOauthUser(event):
@@ -72,5 +41,3 @@ class OAuthPlugin(GirderPlugin):
         events.bind('no_password_login_attempt', 'oauth', checkOauthUser)
 
         info['apiRoot'].oauth = rest.OAuth()
-
-        SettingDefault.defaults[constants.PluginSettings.PROVIDERS_ENABLED] = []
