+++ plugins/oauth/girder_oauth/providers/base.py
@@ -9,7 +9,8 @@ from girder.models.setting import Setting
 from girder.models.user import User
 from girder.settings import SettingKey
 from girder.utility import config
-from ..constants import PluginSettings
+
+from ..settings import PluginSettings
 
 
 class ProviderBase(object):
