+++ plugins/oauth/girder_oauth/providers/box.py
@@ -4,8 +4,9 @@ from six.moves import urllib
 from girder.api.rest import getApiUrl
 from girder.exceptions import RestException
 from girder.models.setting import Setting
+
 from .base import ProviderBase
-from .. import constants
+from ..settings import PluginSettings
 
 
 class Box(ProviderBase):
@@ -14,16 +15,16 @@ class Box(ProviderBase):
     _API_USER_URL = 'https://api.box.com/2.0/users/me'
 
     def getClientIdSetting(self):
-        return Setting().get(constants.PluginSettings.BOX_CLIENT_ID)
+        return Setting().get(PluginSettings.BOX_CLIENT_ID)
 
     def getClientSecretSetting(self):
-        return Setting().get(constants.PluginSettings.BOX_CLIENT_SECRET)
+        return Setting().get(PluginSettings.BOX_CLIENT_SECRET)
 
     @classmethod
     def getUrl(cls, state):
-        clientId = Setting().get(constants.PluginSettings.BOX_CLIENT_ID)
+        clientId = Setting().get(PluginSettings.BOX_CLIENT_ID)
 
-        if clientId is None:
+        if not clientId:
             raise Exception('No Box client ID setting is present.')
 
         callbackUrl = '/'.join((getApiUrl(), 'oauth', 'box', 'callback'))
