--- plugins/user_quota/girder_user_quota/constants.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-# Constants representing the setting keys for this plugin
-class PluginSettings(object):
-    QUOTA_DEFAULT_USER_QUOTA = 'user_quota.default_user_quota'
-    QUOTA_DEFAULT_COLLECTION_QUOTA = 'user_quota.default_collection_quota'
