+++ scripts/publicNames.txt
@@ -980,8 +980,10 @@ plugins
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
+            settings
+                PluginSettings
+                    AUTOJOIN
             userCreated
-            validateSettings
     dicom_viewer
         girder_dicom_viewer
             DicomItem
@@ -1003,23 +1005,23 @@ plugins
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
-            constants
-                PluginSettings
-                    GOOGLE_ANALYTICS_TRACKING_ID
             rest
                 GoogleAnalytics
                     getId
-            validateTrackingId
+            settings
+                PluginSettings
+                    TRACKING_ID
     gravatar
         girder_gravatar
             GravatarPlugin
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
-            PluginSettings
-                DEFAULT_IMAGE
             computeBaseUrl
             getGravatar
+            settings
+                PluginSettings
+                    DEFAULT_IMAGE
     hashsum_download
         girder_hashsum_download
             HashedFile
@@ -1032,9 +1034,10 @@ plugins
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
-            PluginSettings
-                AUTO_COMPUTE
             SUPPORTED_ALGORITHMS
+            settings
+                PluginSettings
+                    AUTO_COMPUTE
     homepage
         girder_homepage
             HomepagePlugin
@@ -1043,35 +1046,33 @@ plugins
                 load
             constants
                 COLLECTION_NAME
+            rest
+                Homepage
+                    getAssets
+                    getSettings
+            settings
                 PluginSettings
                     HEADER
                     LOGO
                     MARKDOWN
                     SUBHEADER
                     WELCOME_TEXT
-            rest
-                Homepage
-                    getAssets
-                    getSettings
     item_licenses
         girder_item_licenses
             ItemLicensesPlugin
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
-            constants
-                PluginSettings
-                    LICENSES
-                PluginSettingsDefaults
-                    defaults
             postItemAfter
             postItemCopyAfter
             putItemAfter
             rest
                 getLicenses
+            settings
+                PluginSettings
+                    LICENSES
             updateItemLicense
             validateItem
-            validateLicenses
             validateString
     jobs
         girder_jobs
@@ -1132,9 +1133,9 @@ plugins
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
-            constants
+            settings
                 PluginSettings
-                    LDAP_SERVERS
+                    SERVERS
     oauth
         girder_oauth
             OAuthPlugin
@@ -1142,22 +1143,6 @@ plugins
                 DISPLAY_NAME
                 load
             checkOauthUser
-            constants
-                PluginSettings
-                    BITBUCKET_CLIENT_ID
-                    BITBUCKET_CLIENT_SECRET
-                    BOX_CLIENT_ID
-                    BOX_CLIENT_SECRET
-                    GITHUB_CLIENT_ID
-                    GITHUB_CLIENT_SECRET
-                    GLOBUS_CLIENT_ID
-                    GLOBUS_CLIENT_SECRET
-                    GOOGLE_CLIENT_ID
-                    GOOGLE_CLIENT_SECRET
-                    IGNORE_REGISTRATION_POLICY
-                    LINKEDIN_CLIENT_ID
-                    LINKEDIN_CLIENT_SECRET
-                    PROVIDERS_ENABLED
             providers
                 addProvider
                 base
@@ -1216,9 +1201,22 @@ plugins
                 OAuth
                     callback
                     listProviders
-            validateIgnoreRegistrationPolicy
-            validateOtherSettings
-            validateProvidersEnabled
+            settings
+                PluginSettings
+                    BITBUCKET_CLIENT_ID
+                    BITBUCKET_CLIENT_SECRET
+                    BOX_CLIENT_ID
+                    BOX_CLIENT_SECRET
+                    GITHUB_CLIENT_ID
+                    GITHUB_CLIENT_SECRET
+                    GLOBUS_CLIENT_ID
+                    GLOBUS_CLIENT_SECRET
+                    GOOGLE_CLIENT_ID
+                    GOOGLE_CLIENT_SECRET
+                    IGNORE_REGISTRATION_POLICY
+                    LINKEDIN_CLIENT_ID
+                    LINKEDIN_CLIENT_SECRET
+                    PROVIDERS_ENABLED
     terms
         girder_terms
             TermsPlugin
@@ -1251,10 +1249,6 @@ plugins
                 CLIENT_SOURCE_PATH
                 DISPLAY_NAME
                 load
-            constants
-                PluginSettings
-                    QUOTA_DEFAULT_COLLECTION_QUOTA
-                    QUOTA_DEFAULT_USER_QUOTA
             quota
                 QUOTA_FIELD
                 QuotaPolicy
@@ -1266,7 +1260,10 @@ plugins
                     setCollectionQuota
                     setUserQuota
                 ValidateSizeQuota
-            validateSettings
+            settings
+                PluginSettings
+                    DEFAULT_COLLECTION_QUOTA
+                    DEFAULT_USER_QUOTA
     virtual_folders
         girder_virtual_folders
             VirtualFoldersPlugin
